backend/requirements.txt
absl-py==2.3.1
altair==4.2.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
entrypoints==0.4
flatbuffers==25.2.10
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.42.1
numpy==1.24.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
referencing==0.35.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.20.1
rsa==4.9.1
semver==3.0.4
six==1.17.0
smmap==5.0.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
Werkzeug==3.0.6
wrapt==1.17.3
zipp==3.20.2